Refactored Interpreter to have TypeMismatch and ArgumentMismatch as separate modules. Created readonly `isDevMode` on Interpreter
